Add files via upload
This program first prompts the user to enter the number of subjects. It then takes the marks obtained in each subject from the user using a loop. The total marks are calculated by summing up the marks of all subjects.
Next, the average percentage is calculated by dividing the total marks by the total possible marks (number of subjects * 100) and multiplying by 100.
Based on the average percentage, a grade is assigned using an if-else ladder.
Finally, the program displays the total marks, average percentage, and the corresponding grade to the user.